# docs/openapi.yaml
openapi: 3.1.0
info:
  title: sms-hub API
  version: 2.0.0
  summary: 瘦终端 + Hub 短信网关的完整 HTTP 契约(v2)
  description: |
    三个通信平面:
      1. **设备 → Hub**:`POST /hook/{token}`(设备主动上报,token 即全局 DEVICE_TOKEN)。
      2. **浏览器 → Hub**:`/api/*`(除 health/login/metrics 外需 `Authorization: Bearer <token>`)。
      3. **Hub → 设备**:`/{token}/pull|send|delete|at|update`(由瘦终端固件提供,路径级 `servers` 覆盖指向设备;
         token 不匹配返回 404 空响应;请求体必须 `ensure_ascii=False`)。

    卡片为中心:`sim_id`(IMSI 派生)是短信/发件/通知/收件筛选的业务主键;`device_mac` 是承载载体。
    当前卡片类 API 的 `sim_id` 缺省时,恰 1 张启用卡自动推断,0 或 >1 张时返回 400。

    与本文对齐的人类可读文档:`docs/guide.md`(设计与接口权威:架构/可靠性/接口契约/数据模型)。
    本文件是机器可读、可校验的契约来源;前端类型应据此生成以杜绝手写漂移。

  contact: {}
  license:
    name: MIT
servers:
  - url: http://{hubHost}
    description: Hub —— 浏览器与设备上报入口
    variables:
      hubHost:
        description: Hub 监听地址:端口(默认 :8025)
        default: 127.0.0.1:8025

# 默认所有 /api/* 需 Bearer;公开端点用 security: [] 覆盖
security:
  - bearerAuth: []
...
paths:
...
  /api/devices/{mac}:
    parameters:
      - $ref: '#/components/parameters/DeviceMac'
    patch:
      tags: [Devices & SIMs]
      summary: 改设备名 / 启用状态
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PatchDeviceBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  device: {$ref: '#/components/schemas/DeviceBrief'}
                required: [ok, device]
        '404':
          $ref: '#/components/responses/Error'
  /api/sims/{sim_id}:
    parameters:
      - $ref: '#/components/parameters/SimIdPath'
    patch:
      tags: [Devices & SIMs]
      summary: 改卡片备注 / 启用状态
      description: '`enabled` 由 false→true 时立即对其当前承载设备补拉一次。'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PatchSimBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  sim: {$ref: '#/components/schemas/SimBrief'}
                required: [ok, sim]
        '404':
          $ref: '#/components/responses/Error'

  # ════════════ Messages ════════════
  /api/messages:
    get:
      tags: [Messages]
      summary: 查询收件(分页/搜索/按卡筛选)
      parameters:
        - $ref: '#/components/parameters/SimIdQuery'
        - name: q
          in: query
          schema: {type: string}
          description: 搜索发件人/正文/别名
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: {type: integer}
                  messages:
                    type: array
                    items: {$ref: '#/components/schemas/Message'}
                  readonly: {type: boolean}
                required: [total, messages, readonly]
        '400':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Messages]
      summary: 批量删除(写墓碑 + best-effort 设备删除)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: {type: integer}
              required: [ids]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  deleted: {type: integer}
                  device_deleted: {type: integer}
                required: [ok, deleted, device_deleted]
        '400':
          $ref: '#/components/responses/Error'
  /api/messages/export:
    get:
      tags: [Messages]
      summary: 导出收件(CSV/JSON 文件下载)
      parameters:
        - $ref: '#/components/parameters/SimIdQuery'
        - name: q
          in: query
          schema: {type: string}
        - name: fmt
          in: query
          schema: {type: string, enum: [csv, json], default: csv}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 20000, default: 5000}
      responses:
        '200':
          description: 文件下载
          headers:
            Content-Disposition:
              schema: {type: string}
              example: attachment; filename="sms-export.csv"
          content:
            text/csv:
              schema: {type: string}
            application/json:
              schema: {type: string}
  /api/messages/{msg_id}:
    parameters:
      - name: msg_id
        in: path
        required: true
        schema: {type: integer}
        description: Hub 全局消息 id
    get:
      tags: [Messages]
      summary: 单条详情(含通知记录)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MessageDetail'}
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Messages]
      summary: 删除单条(写墓碑 + best-effort 设备删除)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  deleted: {type: integer, description: '恒为 1', example: 1}
                  device_deleted: {type: integer}
                required: [ok, deleted, device_deleted]

  # ════════════ Outbound ════════════
  /api/send:
    post:
      tags: [Outbound]
      summary: 入站发送(写 outbound 队列,sender worker 执行;不故障转移)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendBody'}
      responses:
        '200':
          description: 已入队
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SendResponse'}
        '400':
          $ref: '#/components/responses/Error'
        '503':
          description: DEVICE_TOKEN 未配置,无法发送
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /api/outbound:
    get:
      tags: [Outbound]
      summary: 发件队列(按卡筛选)
      parameters:
        - name: sim_id
          in: query
          required: false
          schema: {type: string}
          description: 具体 sim_id / `online`(仅在线卡) / `all`;缺省=全部(不自动推断、不报 400)
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  outbound:
                    type: array
                    items: {$ref: '#/components/schemas/Outbound'}
                required: [outbound]
  /api/outbound/{ob_id}:
    parameters:
      - name: ob_id
        in: path
        required: true
        schema: {type: integer}
    delete:
      tags: [Outbound]
      summary: 删除发件记录
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ok'}
        '404':
          $ref: '#/components/responses/Error'

  # ════════════ Contacts ════════════
...
  /api/contacts/{phone}:
    parameters:
      - name: phone
        in: path
        required: true
        schema: {type: string}
        description: 发件人号码(canonicalize 归一化)
    put:
      tags: [Contacts]
      summary: 保存别名(空别名则删除)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alias: {type: string}
              required: [alias]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  phone: {type: string}
                  alias: {type: string}
                  deleted: {type: boolean, description: 空别名删除时为 true}
                required: [ok]
        '400':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Contacts]
      summary: 删除别名
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ok'}

  # ════════════ Status & Control ════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 会话令牌(POST /api/login 获取),TTL 30 天。
  parameters:
    SimIdQuery:
      name: sim_id
      in: query
      required: false
      schema: {type: string}
      description: '具体 sim_id / `all`(全部卡) / `online`(仅在线卡);缺省时单卡自动推断,多卡须显式(否则 400)'
    SimIdPath:
      name: sim_id
      in: path
      required: true
      schema: {type: string}
    DeviceMac:
      name: mac
      in: path
      required: true
      schema: {type: string}
      description: 规范化小写无分隔 12 位 hex
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Error:
      type: object
      properties:
        detail: {type: string}
      required: [detail]
    Ok:
      type: object
      properties:
        ok: {type: boolean}
      required: [ok]

    # ── 设备上报 ──
...
    DeviceBrief:
      type: object
      properties:
        mac: {type: string}
        name: {type: string}
        enabled: {type: boolean}
      required: [mac, name, enabled]
...
    SimBrief:
      type: object
      properties:
        sim_id: {type: string}
        name: {type: string}
        enabled: {type: boolean}
      required: [sim_id, name, enabled]

    # ── 请求体 ──
    PatchDeviceBody:
      type: object
      properties:
        name: {type: string, nullable: true}
        enabled: {type: boolean, nullable: true}
    PatchSimBody:
      type: object
      properties:
        name: {type: string, nullable: true}
        enabled: {type: boolean, nullable: true}
...
    SendBody:
      type: object
      required: [to, text]
      properties:
        sim_id: {type: string, nullable: true}
        to: {type: string}
        text: {type: string}
...
    Message:
      type: object
      required: [id, sim_id, sim_name, device_mac, device_msg_id, gw_epoch, sender, text, received_at, code]
      properties:
        id: {type: integer, description: Hub 全局 id}
        sim_id: {type: string}
        sim_name: {type: string}
        device_mac: {type: string}
        device_msg_id: {type: integer}
        gw_epoch: {type: integer}
        sender: {type: string}
        sender_alias: {type: string, nullable: true}
        text: {type: string}
        scts: {type: string, nullable: true, description: 'YYMMDDhhmmsszz 14 位'}
        received_at: {type: string}
        code: {type: string, nullable: true, description: extract_code 提取的验证码}
        complete: {type: integer, description: '0/1'}
        truncated: {type: integer, description: '0/1'}
        blocked: {type: integer, description: '0/1(命中黑名单)'}
        created_at: {type: string}
        brand: {type: string, nullable: true, description: extract_brand 提取的品牌}
    NotifyJob:
      type: object
      properties:
        channel: {type: string}
        status: {type: string, enum: [pending, ok, retry, give_up]}
        attempts: {type: integer}
        last_error: {type: string}
        created_at: {type: string}
      required: [channel, status, attempts, last_error]
    MessageDetail:
      allOf:
        - $ref: '#/components/schemas/Message'
        - type: object
          properties:
            notify:
              type: array
              items: {$ref: '#/components/schemas/NotifyJob'}
          required: [notify]

    # ── 发件 ──
    Outbound:
      type: object
      required: [id, sim_id, device_mac, to_phone, text, device_msg_id, status, parts, attempts, next_attempt_ts, source, last_error, created_at]
      properties:
        id: {type: integer}
        sim_id: {type: string}
        device_mac: {type: string}
        to_phone: {type: string}
        text: {type: string}
        device_msg_id: {type: integer, nullable: true}
        status: {type: string, enum: [pending, retry, sent, give_up]}
        parts: {type: integer}
        attempts: {type: integer}
        next_attempt_ts: {type: number}
        source: {type: string, description: 'webui / admin_relay / telegram'}
        last_error: {type: string}
        created_at: {type: string}
    SendResponse:
      type: object
      required: [ok, queued, id, parts, sim_id, device_mac]
      properties:
        ok: {type: boolean}
        queued: {type: boolean}
        id: {type: integer, description: outbound 行 id}
        parts: {type: integer}
        sim_id: {type: string}
        device_mac: {type: string}
...